+++ lib/torrone/export_manager.rb
@@ -5,7 +5,7 @@ module Torrone
       def export(jasper_print, format = :pdf)
         raise 'Invalid Format' unless VALID_FORMATS.include?(format)
 
-        klass = "Torrone::Exporters::#{format.to_s.capitalize}Exporter".constantize
+        klass = "Torrone::Exporters::#{format.to_s.capitalize}".constantize
         klass.new.export(jasper_print)
       end
     end
